Implemented random initial positions of particles

+++ src/tree.c
@@ -5,7 +5,7 @@ void buildTree() {
     fprintf(stderr, "Implement 'buildTree'\n");
 }
 
-int findNGB(const int ipart, const float hsml, int *ngblist) {
+int findNGB(const int ipart, const double hsml, int *ngblist) {
     fprintf(stderr, "Implement 'findNGB'\n");
     return 0;
 }
